+++ CHANGELOG.md
@@ -5,6 +5,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## [Unreleased] -->
 
+## [2.0.1] - 2022-07-25
+
+### Changed
+
+- Pop list on`shit+click` even if there is only one blend in folder (allow to pop new instance quickly)
+- Add compatibility with Mac for new blender instance
 ## [2.0.0] - 2022-06-30
 
 ### Added